HTML 1st Draft Finished
Added all HTML tags, Started styling and laying everything out. Still need to arrnange and resize images and position text correctly.

May need to rework the HTML to be able to manipulate some of the attributes but haven't gotten to that phase just yet. Will see if I can get it to work with the way it is laid out currently.